Skip the context within the class header.
If we skip the context naively by skipping to the next =>, then if the
class has no context in its declaration but has methods with contexts,
we are going to skip too far.  Instead, first split off the header of
the class (the part between "class" and "where" if it is present), and
skip to the next => within the header.

This should fix https://github.com/elaforge/fast-tags/issues/3.