Resurrect the function 'breakString'.
It was replaced by 'skipString' in commit
fb3e4b11f6024aeae88b4044193373268f415bb4.

The rationale was that tagging files with giant string literals would
take forever.  I'm going to ignore this problem for the time being as
it is not clear how it can be mitigated.  On the one hand, we need to
parse the string as it is going to become part of the prefixes of the
subsequent tokens.  On the other hand, it is very unlikely that a tag
definition follows a literal string (it is allowed in Haskell to write

foo = "foo"; bar = "bar"

but how many people do that?), so that we could ignore all the tokens
between the beginning of the string and the end of the line.